requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.28
python-dateutil==2.8.2
pandas==2.2.0
xlsxwriter==3.1.9
waitress==3.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1 